refactor sgx filenames
