+++ include/sgx/native_symbols_wrapper.h
