+++ src/sgx/SGXWAMRWasmModule.cpp
@@ -2,8 +2,8 @@
 
 #include <faabric/util/func.h>
 #include <sgx/SGXWAMRWasmModule.h>
-#include <sgx/faasm_sgx_attestation.h>
-#include <sgx/faasm_sgx_system.h>
+#include <sgx/attestation.h>
+#include <sgx/system.h>
 #include <wasm/WasmExecutionContext.h>
 
 extern "C"
