+++ src/sgx/enclave.cpp
@@ -1,7 +1,7 @@
 #include <sgx/enclaveConfig.h>
-#include <sgx/faasm_sgx_enclave_types.h>
-#include <sgx/faasm_sgx_error.h>
-#include <sgx/faasm_sgx_native_symbols_wrapper.h>
+#include <sgx/enclave_types.h>
+#include <sgx/error.h>
+#include <sgx/native_symbols_wrapper.h>
 #include <sgx/rw_lock.h>
 
 #include <iwasm/include/wasm_export.h>
@@ -14,7 +14,7 @@
 #include <iwasm/aot/aot_runtime.h>
 
 #if (FAASM_SGX_ATTESTATION)
-#include <sgx/faasm_sgx_attestation.h>
+#include <sgx/attestation.h>
 #endif
 
 #define WASM_CTORS_FUNC_NAME "__wasm_call_ctors"
