+++ src/sgx/enclave.edl
@@ -6,12 +6,12 @@
 
 enclave{
 #if(FAASM_DOCKER)
-    include "/usr/local/code/faasm/include/sgx/faasm_sgx_error.h"
-    include "/usr/local/code/faasm/include/sgx/faasm_sgx_attestation.h"
+    include "/usr/local/code/faasm/include/sgx/error.h"
+    include "/usr/local/code/faasm/include/sgx/attestation.h"
     include "/build/faasm/_deps/wamr_ext-src/core/iwasm/include/wasm_export.h"
 #else
-    include "../../../include/sgx/faasm_sgx_error.h"
-    include "../../../include/sgx/faasm_sgx_attestation.h"
+    include "../../../include/sgx/error.h"
+    include "../../../include/sgx/attestation.h"
     include "../../_deps/wamr_ext-src/core/iwasm/include/wasm_export.h"
 #endif
     from "sgx_tstdc.edl" import *;
