+++ src/sgx/native_symbols.cpp
